@@ -13,8 +13,15 @@ class EventsController < ApplicationController
   end
   
   
-  def new
+  def new 
     @event = Event.new(:endtime => 1.hour.from_now, :period => "Non ripetere")
+    @meeting = @event.build_meeting
+    @place = @meeting.build_place(:address => "Bologna")
+  end
+  
+  def edit
+    @meeting = @event.meeting
+    @place = @meeting.place if @meeting
   end
   
   def create
@@ -67,8 +74,7 @@ class EventsController < ApplicationController
     end    
   end
   
-  def edit
-  end
+  
   
   def update
     @event = Event.find_by_id(params[:event][:id])
