@@ -1,9 +1,9 @@
 class Place < ActiveRecord::Base
-  belongs_to :comunes, :class_name => 'Comune', :foreign_key => :comune_id
-  has_many :users, :class_name => 'User'
-  has_many :users, :class_name => 'User'
-  has_many :meetings, :class_name => 'Meeting'
+  belongs_to :comune, :class_name => 'Comune', :foreign_key => :comune_id
+  #has_many :users, :class_name => 'User'
+  #has_many :users, :class_name => 'User'
+  has_one :meeting, :class_name => 'Meeting'
   
-  has_many :events, :through => :meetings, :class_name => 'Event'
+#  belongs_to :event, :class_name => 'Event'
 
 end
