@@ -2,24 +2,32 @@ require File.expand_path('../boot', __FILE__)
 
 require 'rails/all'
 
+
 require 'fb_graph'
 
-# If you have a Gemfile, require the gems listed there, including any gems
-# you've limited to :test, :development, or :production.
-Bundler.require(:default, Rails.env) if defined?(Bundler)
+if defined?(Bundler)
+  # If you precompile assets before deploying to production, use this line
+  Bundler.require *Rails.groups(:assets => %w(development test))
+  # If you want your assets lazily compiled in production, use this line
+  # Bundler.require(:default, :assets, Rails.env)
+end
 
 module DemocracyOnline3
   class Application < Rails::Application
-    #config.active_record.observers = :user_observer
-    config.encoding = "UTF-8"
-    config.coding = "UTF-8"
+    config.encoding = "utf-8"
+    config.coding = "utf-8"
 
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
+
+    # Enable the asset pipeline
+    config.assets.enabled = true
+
+    # Version of your assets, change this if you want to expire all your assets
+    config.assets.version = '1.0'
     
     config.autoload_paths << "#{Rails.root}/lib"
     config.time_zone = 'UTC' 
     config.i18n.default_locale = :it 
-
   end
 end
