@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-DemocracyOnline3::Application.config.secret_token = '0d9f5a37f8060a18e7d884629728bcaaf2fc78bd30d34b126a307c5e84dab16292d083af8631e1b717272dca251287f5bef9177f1c03e37f2df5f13a3826c4fc'
+DemocracyOnline3::Application.config.secret_token = '965222823d6d54b3851c58110e6e3eb524a280c1ec41c217769a47b611923513242065e053c111d7ce6efd6dc67699a66f16aca20c01aac2a1b29b1eae3865b4'
